breaking out grouped track table
